+++ frontend/public/changelog.md
@@ -7,6 +7,7 @@
 - Menambahkan daftar frasa pencarian populer di beranda
 - Menawarkan entri glosarium yang mirip ketika frasa yang dicari tidak ditemukan
 - Menambahkan info cara main [Susun Kata](/gim/susun-kata)
+- Menambahkan lencana sinonim dan antonim di [Tesaurus](/tesaurus)
 
 ## 2026-03-01 Minggu